+++ example/index.php
@@ -19,7 +19,8 @@ require_once __DIR__ .'/../vendor/autoload.php';
 //$response = parcel_track()->postLaju()->setTrackingNumber("ER287051644MY")->fetch();
 //$response = parcel_track()->lelExpress()->setTrackingNumber("MYMP000000573505")->fetch();
 //$response = parcel_track()->dhlECommerce()->setTrackingNumber("5218031053514008AAAA")->fetch();
-$response = parcel_track()->ktmd()->setTrackingNumber("103154269")->fetch();
+//$response = parcel_track()->ktmd()->setTrackingNumber("103154269")->fetch();
+$response = parcel_track()->ups()->setTrackingNumber("1Z0V255F0498628539")->fetch();
 
 //$response = parcel_track()->setTrackingNumber("EZP843055940197")->checkCourier();